requirements.txt
# YOLOv5 requirements
# Usage: pip install -r requirements.txt

# Base ----------------------------------------
matplotlib==3.5.2
numpy==1.21.6
opencv-python==4.5.5.64
Pillow==9.5.0
PyYAML==6.0.1
ruamel.yaml==0.17.4
requests==2.27.1
scipy==1.5.4
torch==1.10.1
torchvision==0.11.2
tqdm==4.64.0
protobuf==3.20.3
flatbuffers==1.12
fast-histogram==0.11

# Logging -------------------------------------
tensorboard==2.8.0
# wandb
# clearml

# Plotting ------------------------------------
pandas==1.4.3
seaborn==0.12.2

# Export --------------------------------------
# coremltools>=5.2  # CoreML export
onnx==1.14.0  # ONNX export
onnxoptimizer==0.2.7
onnxruntime==1.14.0
# onnx-simplifier>=0.4.1  # ONNX simplifier
# nvidia-pyindex  # TensorRT export
# nvidia-tensorrt  # TensorRT export
# scikit-learn==0.19.2  # CoreML quantization
tensorflow==2.8.0  # TFLite export (or tensorflow-cpu, tensorflow-aarch64)
# tensorflowjs>=3.9.0  # TF.js export
# openvino-dev  # OpenVINO export

# Extras --------------------------------------
#ipython  # interactive notebook
psutil==5.9.0  # system utilization
thop>=0.1.1  # FLOPs computation
# albumentations>=1.0.3
# pycocotools>=2.0  # COCO mAP
# roboflow

#SingTown
watchdog==2.2.1
oss2==2.16.0